[caching] Move network statistics functions into separate file
